Fixed GCC warnings: -Wconversion -Wshadow -Wunused-parameter. JIRA:CXX-5413
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@74404 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/serial/datatool/traversal_pattern_match_callback.cpp
@@ -229,7 +229,7 @@ CRef<CTraversalNode> CTraversalPatternMatchCallback::x_TranslateArgToNode(
     const CTraversalSpecFileParser::TPattern &extra_arg_pattern )
 {
     // how many levels do we have to climb up to find the extra_arg_pattern?
-    const int levels_to_climb = ( main_pattern.size() - extra_arg_pattern.size() );
+    const int levels_to_climb = (int)(main_pattern.size() - extra_arg_pattern.size());
 
     // climb up
     int level_up = 0;
